move gas related contents to a folder
